Bump to 0.1.16, update readme, move to github actions, and add bitcoin integration testing

+++ programs/lunar/crates/lunar-core/src/lib.rs
@@ -49,15 +49,15 @@ pub enum EntryData {
     Invoice {
         wallet: Wallet,
         currency: Currency,
-        amount: u64,
+        amount: f64,
         user_data: Option<Vec<u8>>,
     },
     Sale {
         wallet: Wallet,
         currency: Currency,
-        amount: u64,
+        amount: f64,
         address: String,
-        confirmations: i64,
+        confirmations: u64,
         user_data: Option<Vec<u8>>,
     },
 }
@@ -74,7 +74,7 @@ pub enum ExitData {
     },
     Sale {
         funded: bool,
-        amount: u64,
+        amount: f64,
         user_data: Option<Vec<u8>>,
     },
 }
